Sun, Feb 2, 2025, 3:09 PM +01:00

+++ credit/credit.c
@@ -1,5 +1,5 @@
-#include <stdio.h>
 #include <cs50.h>
+#include <stdio.h>
 
 long exponential(int n);
 
@@ -7,7 +7,12 @@ int main(void)
 {
     int length = 0;
 
+    // Long to avoid overflow problem
     long number = get_long("Card number: ");
+
+    // Length of the number
+    // If I divide by 15 and the number is not equal to zero than length is 16, otherwhise
+    // I try with smaller numbers
     if ((number / exponential(15)) != 0)
     {
         length = 16;
@@ -24,6 +29,7 @@ int main(void)
     {
         length = 13;
     }
+    // If it's less than 13 then is invalid
     else
     {
         printf("INVALID\n");
@@ -31,34 +37,41 @@ int main(void)
 
     int y = 1;
     int sum = 0;
-    long digit;
+    int digit;
     long digits;
     int doubled;
 
     for (int j = 0; j < length; j++)
     {
+        // To isolate the digit. I obtain the number from the module and then
+        // I divide that number by 10 and the result is the digit (since is an int
+        // the decimal will be truncated)
         digits = (number % exponential(y));
         digit = (digits / exponential(y - 1));
 
-        //Se y è dispari sommo la cifra, se y è pari allora moltiflico per due la cifra prima di sommarla
+        // If y is even then the program multiply the digiit by two
+        //  and the adds the digit to the sum
         if (y % 2 == 0)
         {
             doubled = digit * 2;
             if (doubled >= 10)
             {
+                // If the digit multiplied by two is greater than 10,
+                // then I have to sum the digits of the program
                 sum = sum + doubled / 10 + doubled % 10;
             }
             else
             {
                 sum += doubled;
             }
+            // If y is odd the program adds the digit to the sum
         }
         else
         {
             sum += digit;
         }
+        // Increment the power to isolate another digit on the left
         y += 1;
-
     }
 
     // Second digit MASTERCARD
@@ -67,6 +80,7 @@ int main(void)
     // Second digit AMEX
     int second_digit_X = (number / exponential(13)) % 10;
 
+    // If the last digit of the sum is != 0 the number is invalid
     if (sum % 10 != 0)
     {
         printf("INVALID\n");
@@ -111,10 +125,9 @@ int main(void)
             }
         }
     }
-
 }
 
-
+// Function to make power of 10 based on the number in input (which is the exponent)
 long exponential(int n)
 {
     long x = 1;
